refactor(web): extract and story status controls
N/A

- Extract chat lifecycle indicators from ChatPage and reusable appearance radio cards/previews from SettingsPage without changing their behavior.
- Add 13 deterministic Storybook states covering sandbox/runner/MCP status bands and appearance selection layouts.
- Keep existing page tests pointed at the extracted components and add a shared chat layout constant for the moved surfaces.

ELI5: the final reusable UI pieces are no longer buried inside giant pages, so Storybook can render and snapshot them directly while the pages keep using the same components.

```text
page-private UI -> reusable leaf component -> focused stories and pinned snapshots
```

- `pnpm --filter web run build:storybook`
- `pnpm --filter web run type-check`
- `pnpm --filter web run lint`
- `pnpm --filter web exec vitest run src/pages/ConnectionIndicator.test.tsx src/pages/RunnerStartingIndicator.test.tsx src/pages/ChatPage.indicators.test.tsx src/pages/ChatPage.mcpStartup.test.ts src/pages/SettingsPage.test.tsx`
- `uv run --no-sync pre-commit run --files $(git ls-files -m -o --exclude-standard)`
- Rendered and reviewed all 13 extracted indicator and appearance stories through the Playwright snapshot harness.
- Generated and reviewed all 13 new Linux baselines from the stacked PR pinned UI Snapshot artifact.

N/A — this is a behavior-preserving extraction plus developer visual-regression coverage.

- [ ] Bug fix
- [ ] Feature
- [ ] UI / frontend change
- [x] Refactor / chore
- [ ] Docs
- [ ] Test / CI
- [ ] Breaking change

- [x] Unit tests added / updated
- [ ] Integration tests added / updated
- [x] E2E tests added / updated
- [x] Manual verification completed
- [ ] Existing tests cover this change
- [ ] Not applicable

Existing ChatPage and SettingsPage tests pass against the extracted components. All new story states were rendered and reviewed locally; pinned Linux baselines are generated in CI.
