+++ web/src/pages/SettingsPage.tsx
@@ -36,13 +36,11 @@ import {
   useEffect,
   useId,
   useMemo,
-  useRef,
   useState,
 } from "react";
 import {
   ArchiveRestoreIcon,
   AlertTriangleIcon,
-  CheckIcon,
   KeyRoundIcon,
   LaptopMinimalIcon,
   LogOutIcon,
@@ -61,6 +59,12 @@ import {
 import { useTheme } from "next-themes";
 import { PageScroll } from "@/components/PageScroll";
 import { ThemeColorPicker } from "@/components/theme/ThemeColorPicker";
+import { CardRadioGroup } from "@/components/theme/CardRadioGroup";
+import {
+  ModePreview,
+  PaletteChip,
+  PaletteSwatchPreview,
+} from "@/components/theme/AppearancePreviews";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Switch } from "@/components/ui/switch";
@@ -164,7 +168,6 @@ import {
   DEFAULT_PALETTE,
   isThemeSelection,
   PALETTES,
-  type PaletteSwatch,
   readThemePalette,
   type ThemeSelection,
   writeThemePalette,
@@ -311,37 +314,6 @@ const workspacePanelCards: {
   { value: "collapsed", label: "Collapsed", icon: PanelRightCloseIcon },
 ];
 
-/**
- * Checkmark badge pinned to the top-right corner of a selected card. Shared by
- * every appearance radiogroup so "selected" reads identically everywhere.
- */
-function SelectedBadge() {
-  return (
-    <span
-      aria-hidden
-      className="absolute right-1.5 top-1.5 flex size-4 items-center justify-center rounded-full bg-primary text-primary-foreground shadow-sm"
-    >
-      <CheckIcon className="size-3" />
-    </span>
-  );
-}
-
-/**
- * Shared card styling for the appearance radiogroups. Selected cards carry the
- * accent border + a subtle accent wash (paired with <SelectedBadge/>); the rest
- * highlight their border and lift on hover. focus-visible keeps the global
- * outline ring, so keyboard focus stays visually distinct from selection.
- */
-function themeCardClass(selected: boolean, layout?: string) {
-  return cn(
-    "relative flex flex-col rounded-lg border-2 transition-[color,background-color,border-color,box-shadow]",
-    selected
-      ? "border-primary bg-primary/5"
-      : "border-border hover:border-border-strong hover:bg-muted hover:shadow-sm",
-    layout,
-  );
-}
-
 /** Centered icon + label body shared by the Mode and Terminal theme cards. */
 function iconCardBody(Icon: typeof SunIcon, label: string) {
   return (
@@ -349,143 +321,6 @@ function iconCardBody(Icon: typeof SunIcon, label: string) {
       <Icon className="size-6 text-muted-foreground" />
       <span className="text-ui font-medium">{label}</span>
     </>
-  );
-}
-
-// Neutral light/dark window tones for the Mode preview tiles. These are about
-// light-vs-dark only (not the color theme), so they stay grayscale.
-const LIGHT_MODE_PREVIEW: PaletteSwatch = {
-  bg: "#e9ebee",
-  card: "#ffffff",
-  accent: "#aab2bd",
-  border: "#d7dbe0",
-  text: "#11171c",
-};
-const DARK_MODE_PREVIEW: PaletteSwatch = {
-  bg: "#0e1013",
-  card: "#232a33",
-  accent: "#5b6672",
-  border: "#2b333d",
-  text: "#e6edf3",
-};
-
-/**
- * Mini app-window mock for a Mode tile, reusing {@link PaletteSwatchPreview}. A
- * light or dark two-pane window; "system" shows one window split diagonally —
- * light on the near side, dark on the far — to signal "follow the OS".
- */
-function ModePreview({ variant }: { variant: ThemeMode }) {
-  if (variant === "light") return <PaletteSwatchPreview swatch={LIGHT_MODE_PREVIEW} />;
-  if (variant === "dark") return <PaletteSwatchPreview swatch={DARK_MODE_PREVIEW} />;
-  return (
-    <div className="relative h-16 w-full">
-      <PaletteSwatchPreview swatch={LIGHT_MODE_PREVIEW} />
-      <div
-        aria-hidden
-        className="absolute inset-0"
-        style={{ clipPath: "polygon(62% 0, 100% 0, 100% 100%, 38% 100%)" }}
-      >
-        <PaletteSwatchPreview swatch={DARK_MODE_PREVIEW} />
-      </div>
-    </div>
-  );
-}
-
-/** Small swatch chip (canvas + accent dot) for the color-theme dropdown. */
-function PaletteChip({ swatch }: { swatch: PaletteSwatch }) {
-  return (
-    <span
-      aria-hidden
-      className="flex size-5 shrink-0 items-center justify-center rounded-md border"
-      style={{ backgroundColor: swatch.bg, borderColor: swatch.border }}
-    >
-      <span className="size-2 rounded-full" style={{ backgroundColor: swatch.accent }} />
-    </span>
-  );
-}
-
-/** One option in a {@link CardRadioGroup}. */
-interface CardRadioOption<T extends string> {
-  value: T;
-  testId: string;
-  body: ReactNode;
-  /** Optional native tooltip (used for the palette blurbs). */
-  title?: string;
-}
-
-/**
- * Accessible card radiogroup shared by all three appearance pickers. Implements
- * the WAI-ARIA radiogroup pattern: a roving tabindex (only the selected card is
- * tabbable), arrow keys move selection within the group, and Enter/Space select
- * the focused card. `labelledBy` points at the subsection heading so the group's
- * accessible name matches its visible label.
- */
-function CardRadioGroup<T extends string>({
-  labelledBy,
-  value,
-  onSelect,
-  componentId,
-  items,
-  className,
-  cardClassName,
-}: {
-  labelledBy: string;
-  value: T;
-  onSelect: (value: T) => void;
-  // Opt-in analytics id for the whole picker. When set, a selection reports the
-  // chosen value to the host sink (see `lib/analytics.ts`). Card values are a
-  // bounded set, so the value is sent. Covers both the click and arrow-key paths.
-  componentId?: string;
-  items: readonly CardRadioOption<T>[];
-  className?: string;
-  cardClassName?: string;
-}) {
-  const { trackValueChange } = useOmnigentAnalytics();
-  const select = componentId
-    ? (next: T) => {
-        trackValueChange(componentId, "select", next, { valueHasNoPii: true });
-        onSelect(next);
-      }
-    : onSelect;
-  // Keep a handle on each card so arrow-key navigation can move focus as it
-  // moves selection (selection-follows-focus, per the radiogroup pattern).
-  const refs = useRef(new Map<T, HTMLButtonElement | null>());
-
-  return (
-    <div role="radiogroup" aria-labelledby={labelledBy} className={className}>
-      {items.map((item, index) => {
-        const selected = item.value === value;
-        return (
-          <button
-            key={item.value}
-            ref={(el) => {
-              refs.current.set(item.value, el);
-            }}
-            type="button"
-            role="radio"
-            aria-checked={selected}
-            tabIndex={selected ? 0 : -1}
-            title={item.title}
-            data-testid={item.testId}
-            onClick={() => select(item.value)}
-            onKeyDown={(event) => {
-              const forward = event.key === "ArrowRight" || event.key === "ArrowDown";
-              const backward = event.key === "ArrowLeft" || event.key === "ArrowUp";
-              if (!forward && !backward) return;
-              event.preventDefault();
-              const nextIndex = (index + (forward ? 1 : -1) + items.length) % items.length;
-              const next = items[nextIndex].value;
-              select(next);
-              refs.current.get(next)?.focus();
-            }}
-            className={themeCardClass(selected, cardClassName)}
-          >
-            {selected && <SelectedBadge />}
-            {item.body}
-          </button>
-        );
-      })}
-    </div>
   );
 }
 
@@ -817,50 +652,6 @@ function ColorThemeControl() {
         </div>
       </div>
     </ThemeSubsection>
-  );
-}
-
-/**
- * Miniature "app window" preview for a palette: a canvas with a small sidebar
- * and content card, a few text lines, and an accent chip — built purely from
- * the swatch colors so each palette reads at a glance.
- */
-function PaletteSwatchPreview({ swatch }: { swatch: PaletteSwatch }) {
-  return (
-    <div
-      aria-hidden
-      className="flex h-16 w-full gap-1.5 overflow-hidden rounded-lg p-1.5"
-      style={{ backgroundColor: swatch.bg, border: `1px solid ${swatch.border}` }}
-    >
-      <div
-        className="flex w-1/3 flex-col gap-1 rounded-md p-1"
-        style={{ backgroundColor: swatch.card, border: `1px solid ${swatch.border}` }}
-      >
-        <div className="size-1.5 rounded-full" style={{ backgroundColor: swatch.accent }} />
-        <div
-          className="h-1 w-4/5 rounded-full"
-          style={{ backgroundColor: swatch.text, opacity: 0.35 }}
-        />
-        <div
-          className="h-1 w-3/5 rounded-full"
-          style={{ backgroundColor: swatch.text, opacity: 0.25 }}
-        />
-      </div>
-      <div
-        className="flex flex-1 flex-col gap-1 rounded-md p-1.5"
-        style={{ backgroundColor: swatch.card, border: `1px solid ${swatch.border}` }}
-      >
-        <div
-          className="h-1 w-3/4 rounded-full"
-          style={{ backgroundColor: swatch.text, opacity: 0.5 }}
-        />
-        <div
-          className="h-1 w-1/2 rounded-full"
-          style={{ backgroundColor: swatch.text, opacity: 0.3 }}
-        />
-        <div className="mt-auto h-2.5 w-2/5 rounded" style={{ backgroundColor: swatch.accent }} />
-      </div>
-    </div>
   );
 }
