+++ web/src/pages/chatLayout.ts
@@ -1,0 +1,2 @@
+export const CHAT_COLUMN_WIDTH =
+  "max-w-3xl min-[1921px]:max-w-4xl min-[2561px]:max-w-[clamp(64rem,40vw,100rem)]";
